Get rid of some atomics
